Enh: split test file
